+++ deltachat/message.go
@@ -112,9 +112,9 @@ func (self *Message) StatusUpdates(lastKnownSerial uint) (string, error) {
 
 // Get info from this webxdc message.
 func (self *Message) WebxdcInfo() (*WebxdcMsgInfo, error) {
-	var info *WebxdcMsgInfo
-	err := self.rpc().CallResult(info, "get_webxdc_info", self.Account.Id, self.Id)
-	return info, err
+	var info WebxdcMsgInfo
+	err := self.rpc().CallResult(&info, "get_webxdc_info", self.Account.Id, self.Id)
+	return &info, err
 }
 
 func (self *Message) rpc() Rpc {