+++ deltachat/message_test.go
@@ -37,6 +37,79 @@ func TestMessage(t *testing.T) {
 	assert.Nil(t, msg.MarkSeen())
 
 	assert.Nil(t, msg.SendReaction(":)"))
+
+	assert.Nil(t, msg.Delete())
+}
+
+func TestMessage_WebxdcInfo(t *testing.T) {
+	acc := acfactory.GetOnlineAccount()
+	defer acc.Manager.Rpc.Stop()
+
+	chat, err := acc.CreateChat(acc)
+	assert.Nil(t, err)
+
+	msg, err := chat.SendText("test")
+	assert.Nil(t, err)
+	info, err := msg.WebxdcInfo()
+	assert.NotNil(t, err)
+
+	msg, err = chat.SendMsg(MsgData{Text: "testing webxdc", File: acfactory.GetTestWebxdc()})
+	assert.Nil(t, err)
+
+	info, err = msg.WebxdcInfo()
+	assert.Nil(t, err)
+	assert.NotEmpty(t, info.Name)
+}
+
+func TestMessage_StatusUpdates(t *testing.T) {
+	acc1 := acfactory.GetOnlineAccount()
+	defer acc1.Manager.Rpc.Stop()
+	acc2 := acfactory.GetOnlineAccount()
+	defer acc2.Manager.Rpc.Stop()
+
+	chat1, err := acc1.CreateChat(acc2)
+	assert.Nil(t, err)
+
+	msg, err := chat1.SendMsg(MsgData{Text: "testing webxdc", File: acfactory.GetTestWebxdc()})
+	assert.Nil(t, err)
+	snapshot, err := acfactory.GetNextMsg(acc2)
+	assert.Nil(t, err)
+
+	assert.Nil(t, msg.SendStatusUpdate(`{"payload": "test payload"}`, "update 1"))
+	WaitForEvent(acc2, EVENT_WEBXDC_STATUS_UPDATE)
+
+	msg = &Message{acc2, snapshot.Id}
+	updates, err := msg.StatusUpdates(0)
+	assert.Nil(t, err)
+	assert.Contains(t, updates, "test payload")
+}
+
+func TestMessage_ContinueAutocryptKeyTransfer(t *testing.T) {
+	acc1 := acfactory.GetOnlineAccount()
+	defer acc1.Manager.Rpc.Stop()
+	acc2 := acfactory.GetUnconfiguredAccount()
+	defer acc2.Manager.Rpc.Stop()
+
+	addr, err := acc1.GetConfig("configured_addr")
+	assert.Nil(t, err)
+	assert.NotEmpty(t, addr)
+	acc2.SetConfig("addr", addr)
+	password, err := acc1.GetConfig("configured_mail_pw")
+	assert.Nil(t, err)
+	assert.NotEmpty(t, password)
+	acc2.SetConfig("mail_pw", password)
+	assert.Nil(t, acc2.Configure())
+
+	code, err := acc1.InitiateAutocryptKeyTransfer()
+	assert.Nil(t, err)
+	assert.NotEmpty(t, code)
+
+	selfchat, err := acc2.Me().CreateChat()
+	assert.Nil(t, err)
+	event := waitForEvent(acc2, EVENT_MSGS_CHANGED, selfchat.Id)
+	assert.NotEmpty(t, event.MsgId)
+	msg := &Message{acc2, event.MsgId}
+	assert.Nil(t, msg.ContinueAutocryptKeyTransfer(code))
 }
 
 func TestMsgSnapshot_ParseMemberAddedRemoved(t *testing.T) {
@@ -71,6 +144,10 @@ func TestMsgSnapshot_ParseMemberAddedRemoved(t *testing.T) {
 			break
 		}
 	}
+	snapshot, err := msg.Snapshot()
+	assert.Nil(t, err)
+	_, _, err = snapshot.ParseMemberAdded()
+	assert.NotNil(t, err)
 
 	// add new member
 	assert.Nil(t, chat1.AddContact(contact2))
@@ -78,9 +155,11 @@ func TestMsgSnapshot_ParseMemberAddedRemoved(t *testing.T) {
 	event := waitForEvent(acc1, EVENT_MSGS_CHANGED, chat1.Id)
 	assert.NotEmpty(t, event.MsgId)
 	msg = &Message{acc1, event.MsgId}
-	snapshot, err := msg.Snapshot()
+	snapshot, err = msg.Snapshot()
 	assert.Nil(t, err)
 	assert.Equal(t, SYSMSG_TYPE_MEMBER_ADDED_TO_GROUP, snapshot.SystemMessageType)
+	_, _, err = snapshot.ParseMemberRemoved()
+	assert.NotNil(t, err)
 	actor, target, err := snapshot.ParseMemberAdded()
 	assert.Nil(t, err)
 	assert.Equal(t, acc1.Me(), actor)
@@ -103,6 +182,8 @@ func TestMsgSnapshot_ParseMemberAddedRemoved(t *testing.T) {
 	snapshot, err = msg.Snapshot()
 	assert.Nil(t, err)
 	assert.Equal(t, SYSMSG_TYPE_MEMBER_REMOVED_FROM_GROUP, snapshot.SystemMessageType)
+	_, _, err = snapshot.ParseMemberAdded()
+	assert.NotNil(t, err)
 	actor, target, err = snapshot.ParseMemberRemoved()
 	assert.Nil(t, err)
 	assert.Equal(t, acc1.Me(), actor)